src/requirements.txt
aiohttp
aioinflux
asyncpg
dateparser
discord.py
humanize
uvloop